Add a new definition of BLOCK_CIPHER_def1, only enabled when FIPS_CIPHER_DISABLE is defined before including evp.h.
This allows us to only bear the cost of calling FIPS_mode()
and the associated locking when we know we the caller is
trying to instantiate a non-FIPS approved cipher.

Essentially all this change does is add a

+    if (FIPS_mode()) { \
+        FIPSerr(ERR_LIB_FIPS, FIPS_R_NON_FIPS_METHOD); \
+        return NULL; \
+    } \

block into the generated definition of the
const EVP_CIPHER *EVP_##cname##_##mode(void) function,
but only if FIPS_CIPHER_DISABLE is defined before
inclusion of this header file.

We can now add:

 define FIPS_CIPHER_DISABLE 1

lines into all ciphers that instantiate themselves
as EVP as a line just before including include/crypto/evp.h,
and this will allow them to be disabled in FIPS mode.

Also adds the needed include lines to include/crypto/evp.h
in order for FIPSerr(ERR_LIB_FIPS, FIPS_R_NON_FIPS_METHOD)
to be correctly defined.

Signed-off-by: Jeremy Allison <jallison@ciq.com>